email_worker/requirements.txt
alembic==1.11.1
amqp==5.1.1
async-timeout==4.0.2
billiard==3.6.4.0
blinker==1.6.2
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coverage==7.2.7
dnspython==2.3.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
Flask==2.3.2
Flask-Cors==3.0.10
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.0
Mako==1.2.4
MarkupSafe==2.1.2
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.1
pluggy==1.0.0
prompt-toolkit==3.0.38
pytest==7.3.1
pytest-cov==4.1.0
python-dotenv==1.0.0
pytz==2023.3
redis==4.5.5
six==1.16.0
SQLAlchemy==2.0.15
tomli==2.0.1
typing_extensions==4.6.3
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.3.4
WTForms==3.0.1
